requirements.txt
PyPDF2==2.12.1  
requests
bs4
pandas
tabula-py
numpy
openpyxl